update: organize component imports by indexing
