+++ src/components/MainContent/About/About.jsx
@@ -1,7 +1,7 @@
 import { faUpRightFromSquare } from "@fortawesome/free-solid-svg-icons";
 import { intersectionService } from "../../../services/intersection-service";
 
-import {Navbar} from "../../Navbar/Navbar";
+import {Navbar} from "../../Navbar";
 
 import "./About.css";
