+++ src/components/MainContent/Contact/index.js
@@ -1,0 +1,1 @@
+export * from './Contact.jsx'
