+++ src/components/MainContent/Hero/index.js
@@ -1,0 +1,1 @@
+export * from './Hero.jsx'
