+++ src/components/MainContent/MainContent.jsx
@@ -5,11 +5,7 @@ import {
   socialMediaService,
 } from "../../services/db-service";
 
-import { About } from "./About/About";
-import { Contact } from "./Contact/Contact";
-import {Hero} from "./Hero/Hero";
-import { Projects } from "./Projects/Projects";
-import { Skills } from "./Skills/Skills";
+import { Hero, About, Contact, Projects, Skills } from "../../components";
 
 function Root({ children }) {
   return <main className="main">{children}</main>;
