+++ src/components/MainContent/Projects/ProjectCard/index.js
@@ -1,0 +1,1 @@
+export * from './ProjectCard.jsx'
