+++ src/components/MainContent/Projects/ProjectFilter/index.js
@@ -1,0 +1,1 @@
+export * from './ProjectFilter.jsx'
