+++ src/components/MainContent/Projects/ProjectModal/index.js
@@ -1,0 +1,1 @@
+export * from './ProjectModal.jsx'
