+++ src/components/MainContent/Projects/Projects.jsx
@@ -1,10 +1,7 @@
 import { faArrowDown } from "@fortawesome/free-solid-svg-icons";
 import { intersectionService } from "../../../services/intersection-service";
 
-import { ProjectModal } from "./ProjectModal/ProjectModal";
-import { ProjectCard } from "./ProjectCard/ProjectCard";
-import { ProjectFilter } from "./ProjectFilter/ProjectFilter";
-import { Modal } from "../../Modal/Modal";
+import { Modal, ProjectModal, ProjectFilter, ProjectCard } from "../../../components";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
 import "./Projects.css";
