+++ src/components/MainContent/Projects/index.js
@@ -1,0 +1,1 @@
+export * from './Projects.jsx'
