+++ src/components/MainContent/Skills/SkillCard/index.js
@@ -1,0 +1,1 @@
+export * from './SkillCard.jsx'
