+++ src/components/MainContent/Skills/SkillModal/index.js
@@ -1,0 +1,1 @@
+export * from './SkillModal.jsx'
