+++ src/components/MainContent/Skills/Skills.jsx
@@ -4,10 +4,8 @@ import {
   faUpRightFromSquare,
 } from "@fortawesome/free-solid-svg-icons";
 
+import { Modal, SkillCard, SkillModal } from '../../../components'
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { Modal } from "../../Modal/Modal";
-import { SkillCard } from "./SkillCard/SkillCard";
-import { SkillModal } from "./SkillModal/SkillModal";
 
 import "./Skills.css";
