+++ src/components/MainContent/Skills/index.js
@@ -1,0 +1,1 @@
+export * from './Skills.jsx'
