+++ src/components/MainContent/index.js
@@ -1,0 +1,1 @@
+export * from './MainContent.jsx'
