+++ src/components/MainHeader/HeaderMenu/index.js
@@ -1,0 +1,1 @@
+export * from './HeaderMenu.jsx'
