+++ src/components/MainHeader/index.js
@@ -1,0 +1,1 @@
+export * from './MainHeader.jsx'
