+++ src/components/ScrollTop/index.js
@@ -1,0 +1,1 @@
+export * from './ScrollTop.jsx'
