+++ src/main.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 
-import {App} from "./components/App/App";
+import {App} from "./components";
 
 import "./reset.css";
 import "./index.css";
